refactor: move all Properties classes from core to spring-core
GateProperties, ScheduleProperties, and ResponseType were only referenced
by EndpointGateProperties (spring-core) and are Spring property binding
concerns, so they belong in the spring-core module.

- Delete GateProperties, ScheduleProperties, ResponseType from core
- Move GatePropertiesTest, SchedulePropertiesTest from core to spring-core
- ScheduleProperties.toSchedule() remains package-private; same-package
  access is preserved after move
- Update additional-spring-configuration-metadata.json type references
