+++ spring/core/src/test/java/net/brightroom/endpointgate/spring/core/autoconfigure/EndpointGateAutoConfigurationEmptyPropertiesTest.java
@@ -3,9 +3,9 @@ package net.brightroom.endpointgate.spring.core.autoconfigure;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Map;
-import net.brightroom.endpointgate.core.properties.ResponseType;
 import net.brightroom.endpointgate.spring.core.properties.EndpointGateProperties;
 import net.brightroom.endpointgate.spring.core.properties.ResponseProperties;
+import net.brightroom.endpointgate.spring.core.properties.ResponseType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
