+++ spring/core/src/test/java/net/brightroom/endpointgate/spring/core/properties/GatePropertiesTest.java
@@ -1,4 +1,4 @@
-package net.brightroom.endpointgate.core.properties;
+package net.brightroom.endpointgate.spring.core.properties;
 
 import static org.junit.jupiter.api.Assertions.*;
